requirements.txt
aiogram~=3.15.0
python-dotenv==1.0.0
supabase~=2.10.0
requests~=2.32.3
pandas~=2.2.3